+++ include/libKitsunemimiSakuraLang/items/item_methods.h
@@ -28,7 +28,7 @@
 
 #include <libKitsunemimiCommon/common_items/data_items.h>
 
-#include <items/sakura_items.h>
+#include <libKitsunemimiSakuraLang/items/sakura_items.h>
 
 namespace Kitsunemimi
 {
