+++ src/blossom.cpp
@@ -22,7 +22,7 @@
 
 #include <libKitsunemimiSakuraLang/blossom.h>
 
-#include <items/item_methods.h>
+#include <libKitsunemimiSakuraLang/items/item_methods.h>
 
 namespace Kitsunemimi
 {
