+++ src/items/sakura_items.cpp
@@ -20,7 +20,7 @@
  *      limitations under the License.
  */
 
-#include "sakura_items.h"
+#include <libKitsunemimiSakuraLang/items/sakura_items.h>
 
 namespace Kitsunemimi
 {
