+++ src/parsing/sakura_parsing.cpp
@@ -22,7 +22,7 @@
 
 #include "sakura_parsing.h"
 
-#include <items/sakura_items.h>
+#include <libKitsunemimiSakuraLang/items/sakura_items.h>
 #include <sakura_garden.h>
 #include <parsing/sakura_parser_interface.h>
