+++ src/sakura_garden.cpp
@@ -21,7 +21,7 @@
  */
 
 #include <sakura_garden.h>
-#include <items/sakura_items.h>
+#include <libKitsunemimiSakuraLang/items/sakura_items.h>
 #include <parsing/sakura_parsing.h>
 
 #include <libKitsunemimiCommon/common_methods/string_methods.h>
