+++ src/validator.cpp
@@ -23,6 +23,7 @@
 #include "validator.h"
 
 #include <libKitsunemimiSakuraLang/blossom.h>
+#include <libKitsunemimiSakuraLang/items/item_methods.h>
 #include <libKitsunemimiSakuraLang/sakura_lang_interface.h>
 
 namespace Kitsunemimi
